Moved dataset table actions to new component
