+++ src/app/datasets/dataset-table-actions/dataset-table-actions.component.spec.ts
@@ -1,0 +1,144 @@
+import {
+  async,
+  ComponentFixture,
+  TestBed,
+  inject
+} from "@angular/core/testing";
+
+import { DatasetTableActionsComponent } from "./dataset-table-actions.component";
+import { MockStore, MockArchivingService } from "shared/MockStubs";
+import { NO_ERRORS_SCHEMA } from "@angular/core";
+import { Store, StoreModule } from "@ngrx/store";
+import { ArchViewMode, Dataset } from "state-management/models";
+import {
+  setArchiveViewModeAction,
+  setPublicViewModeAction,
+  addToBatchAction,
+  clearSelectionAction
+} from "state-management/actions/datasets.actions";
+import { AppConfigModule, APP_CONFIG } from "app-config.module";
+import { MatDialogModule } from "@angular/material";
+import { ArchivingService } from "datasets/archiving.service";
+
+describe("DatasetTableActionsComponent", () => {
+  let component: DatasetTableActionsComponent;
+  let fixture: ComponentFixture<DatasetTableActionsComponent>;
+
+  let store: MockStore;
+  let dispatchSpy;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      schemas: [NO_ERRORS_SCHEMA],
+      declarations: [DatasetTableActionsComponent],
+      imports: [AppConfigModule, MatDialogModule, StoreModule.forRoot({})]
+    });
+    TestBed.overrideComponent(DatasetTableActionsComponent, {
+      set: {
+        providers: [
+          { provide: APP_CONFIG, useValue: { archiveWorkflowEnabled: true } },
+          { provide: ArchivingService, useClass: MockArchivingService }
+        ]
+      }
+    });
+    TestBed.compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(DatasetTableActionsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  beforeEach(inject([Store], (mockStore: MockStore) => {
+    store = mockStore;
+  }));
+
+  afterEach(() => {
+    fixture.destroy();
+  });
+
+  it("should create", () => {
+    expect(component).toBeTruthy();
+  });
+
+  it("should contain mode switching buttons", () => {
+    const compiled = fixture.debugElement.nativeElement;
+    expect(compiled.querySelector(".archivable")).toBeTruthy();
+    expect(compiled.querySelector(".archivable").textContent).toContain(
+      "Archivable"
+    );
+    expect(compiled.querySelector(".retrievable")).toBeTruthy();
+    expect(compiled.querySelector(".retrievable").textContent).toContain(
+      "Retrievable"
+    );
+    expect(compiled.querySelector(".all")).toBeTruthy();
+    expect(compiled.querySelector(".all").textContent).toContain("All");
+  });
+
+  describe("#onModeChange()", () => {
+    it("should dispatch a SetViewModeAction", () => {
+      dispatchSpy = spyOn(store, "dispatch");
+
+      const event = "test";
+      const modeToggle = ArchViewMode.all;
+
+      component.onModeChange(event, modeToggle);
+
+      expect(dispatchSpy).toHaveBeenCalledTimes(1);
+      expect(dispatchSpy).toHaveBeenCalledWith(
+        setArchiveViewModeAction({ modeToggle })
+      );
+    });
+  });
+
+  describe("#onViewPublicChange()", () => {
+    it("should dispatch a SetPublicViewModeAction", () => {
+      dispatchSpy = spyOn(store, "dispatch");
+
+      const viewPublic = false;
+      component.onViewPublicChange(viewPublic);
+
+      expect(dispatchSpy).toHaveBeenCalledTimes(1);
+      expect(dispatchSpy).toHaveBeenCalledWith(
+        setPublicViewModeAction({ isPublished: viewPublic })
+      );
+    });
+  });
+
+  describe("#isEmptySelection()", () => {
+    it("should return true if the length of selectedSets equals 0", () => {
+      const isEmpty = component.isEmptySelection();
+
+      expect(isEmpty).toEqual(true);
+    });
+
+    it("should return false if the length of selectedSets is larger than 0", () => {
+      component.selectedSets = [new Dataset()];
+
+      const isEmpty = component.isEmptySelection();
+
+      expect(isEmpty).toEqual(false);
+    });
+  });
+
+  describe("#archiveClickHandle()", () => {
+    xit("should...", () => {});
+  });
+
+  describe("#retrieveClickHandle()", () => {
+    xit("should...", () => {});
+  });
+
+  describe("#onAddToBatch()", () => {
+    it("should dispatch an addToBatchAction and a clearSelectionAction", () => {
+      dispatchSpy = spyOn(store, "dispatch");
+
+      component.onAddToBatch();
+
+      expect(dispatchSpy).toHaveBeenCalledTimes(2);
+      expect(dispatchSpy).toHaveBeenCalledWith(addToBatchAction());
+      expect(dispatchSpy).toHaveBeenCalledWith(clearSelectionAction());
+    });
+  });
+});
