+++ src/app/datasets/dataset-table/dataset-table.component.spec.ts
@@ -1,19 +1,10 @@
 import { APP_CONFIG, AppConfigModule } from "app-config.module";
-import { ArchivingService } from "../archiving.service";
 import {
   DatasetTableComponent,
   SortChangeEvent
 } from "./dataset-table.component";
-import {
-  MatDialogModule,
-  MatTableModule,
-  MatCheckboxChange
-} from "@angular/material";
-import {
-  MockDatasetApi,
-  MockArchivingService,
-  MockStore
-} from "shared/MockStubs";
+import { MatTableModule, MatCheckboxChange } from "@angular/material";
+import { MockStore, MockDatasetApi } from "shared/MockStubs";
 import { NO_ERRORS_SCHEMA } from "@angular/core";
 import { Router } from "@angular/router";
 import {
@@ -22,29 +13,24 @@ import {
   TestBed,
   inject
 } from "@angular/core/testing";
-import { combineReducers, StoreModule, Store } from "@ngrx/store";
-import { datasetsReducer } from "state-management/reducers/datasets.reducer";
-import { jobsReducer } from "state-management/reducers/jobs.reducer";
-import { DatasetApi, Dataset } from "shared/sdk";
+import { StoreModule, Store } from "@ngrx/store";
+import { Dataset, DatasetApi } from "shared/sdk";
 import { SharedCatanieModule } from "shared/shared.module";
-import { ArchViewMode } from "state-management/models";
 import {
-  setArchiveViewModeAction,
-  setPublicViewModeAction,
   selectDatasetAction,
   deselectDatasetAction,
   selectAllDatasetsAction,
   clearSelectionAction,
   changePageAction,
-  sortByColumnAction,
-  addToBatchAction
+  sortByColumnAction
 } from "state-management/actions/datasets.actions";
 import { PageChangeEvent } from "shared/modules/table/table.component";
 import { provideMockStore } from "@ngrx/store/testing";
+import { getDatasets } from "state-management/selectors/datasets.selectors";
 import {
-  getDatasets,
-  getSelectedDatasets
-} from "state-management/selectors/datasets.selectors";
+  selectColumnAction,
+  deselectColumnAction
+} from "state-management/actions/user.actions";
 
 describe("DatasetTableComponent", () => {
   let component: DatasetTableComponent;
@@ -61,22 +47,13 @@ describe("DatasetTableComponent", () => {
       schemas: [NO_ERRORS_SCHEMA],
       imports: [
         AppConfigModule,
-        MatDialogModule,
         MatTableModule,
         SharedCatanieModule,
-        StoreModule.forRoot({
-          datasets: datasetsReducer,
-          root: combineReducers({
-            jobs: jobsReducer
-          })
-        })
+        StoreModule.forRoot({})
       ],
       providers: [
         provideMockStore({
-          selectors: [
-            { selector: getDatasets, value: [] },
-            { selector: getSelectedDatasets, value: [] }
-          ]
+          selectors: [{ selector: getDatasets, value: [] }]
         })
       ],
       declarations: [DatasetTableComponent]
@@ -84,16 +61,12 @@ describe("DatasetTableComponent", () => {
     TestBed.overrideComponent(DatasetTableComponent, {
       set: {
         providers: [
-          { provide: Router, useValue: router },
-          { provide: DatasetApi, useClass: MockDatasetApi },
           {
             provide: APP_CONFIG,
-            useValue: {
-              disabledDatasetColumns: [],
-              archiveWorkflowEnabled: true
-            }
+            useValue: {}
           },
-          { provide: ArchivingService, useClass: MockArchivingService }
+          { provide: DatasetApi, useClass: MockDatasetApi },
+          { provide: Router, useValue: router }
         ]
       }
     });
@@ -118,20 +91,6 @@ describe("DatasetTableComponent", () => {
     expect(component).toBeTruthy();
   });
 
-  it("should contain mode switching buttons", () => {
-    const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector(".archivable")).toBeTruthy();
-    expect(compiled.querySelector(".archivable").textContent).toContain(
-      "Archivable"
-    );
-    expect(compiled.querySelector(".retrievable")).toBeTruthy();
-    expect(compiled.querySelector(".retrievable").textContent).toContain(
-      "Retrievable"
-    );
-    expect(compiled.querySelector(".all")).toBeTruthy();
-    expect(compiled.querySelector(".all").textContent).toContain("All");
-  });
-
   describe("#doSettingsClick()", () => {
     it("should emit a MouseEvent on click", () => {
       const emitSpy = spyOn(component.settingsClick, "emit");
@@ -142,44 +101,6 @@ describe("DatasetTableComponent", () => {
       expect(emitSpy).toHaveBeenCalledTimes(1);
       expect(emitSpy).toHaveBeenCalledWith(event);
     });
-  });
-
-  describe("#onModeChange()", () => {
-    it("should dispatch a SetViewModeAction", () => {
-      dispatchSpy = spyOn(store, "dispatch");
-
-      const event = "test";
-      const modeToggle = ArchViewMode.all;
-
-      component.onModeChange(event, modeToggle);
-
-      expect(dispatchSpy).toHaveBeenCalledTimes(1);
-      expect(dispatchSpy).toHaveBeenCalledWith(
-        setArchiveViewModeAction({ modeToggle })
-      );
-    });
-  });
-
-  describe("#onViewPublicChange()", () => {
-    it("should dispatch a SetPublicViewModeAction", () => {
-      dispatchSpy = spyOn(store, "dispatch");
-
-      const viewPublic = false;
-      component.onViewPublicChange(viewPublic);
-
-      expect(dispatchSpy).toHaveBeenCalledTimes(1);
-      expect(dispatchSpy).toHaveBeenCalledWith(
-        setPublicViewModeAction({ isPublished: viewPublic })
-      );
-    });
-  });
-
-  describe("#archiveClickHandle()", () => {
-    xit("should...", () => {});
-  });
-
-  describe("#retrieveClickHandle()", () => {
-    xit("should...", () => {});
   });
 
   describe("#wipCondition()", () => {
@@ -471,7 +392,9 @@ describe("DatasetTableComponent", () => {
   });
 
   describe("#onPageChange()", () => {
-    it("should dispatch a changePangeAction", () => {
+    const column = "image";
+
+    it("should dispatch a changePangeAction and a selectColumnAction if pageSize is less than 50", () => {
       dispatchSpy = spyOn(store, "dispatch");
 
       const event: PageChangeEvent = {
@@ -481,9 +404,29 @@ describe("DatasetTableComponent", () => {
       };
       component.onPageChange(event);
 
-      expect(dispatchSpy).toHaveBeenCalledTimes(1);
+      expect(dispatchSpy).toHaveBeenCalledTimes(2);
       expect(dispatchSpy).toHaveBeenCalledWith(
         changePageAction({ page: event.pageIndex, limit: event.pageSize })
+      );
+      expect(dispatchSpy).toHaveBeenCalledWith(selectColumnAction({ column }));
+    });
+
+    it("should dispatch a changePangeAction and a deselectColumnAction if pageSize is larger than or equal to 50", () => {
+      dispatchSpy = spyOn(store, "dispatch");
+
+      const event: PageChangeEvent = {
+        pageIndex: 0,
+        pageSize: 50,
+        length: 25
+      };
+      component.onPageChange(event);
+
+      expect(dispatchSpy).toHaveBeenCalledTimes(2);
+      expect(dispatchSpy).toHaveBeenCalledWith(
+        changePageAction({ page: event.pageIndex, limit: event.pageSize })
+      );
+      expect(dispatchSpy).toHaveBeenCalledWith(
+        deselectColumnAction({ column })
       );
     });
   });
@@ -502,18 +445,6 @@ describe("DatasetTableComponent", () => {
       expect(dispatchSpy).toHaveBeenCalledWith(
         sortByColumnAction({ column: event.active, direction: event.direction })
       );
-    });
-  });
-
-  describe("#onAddToBatch()", () => {
-    it("should dispatch an addToBatchAction and a clearSelectionAction", () => {
-      dispatchSpy = spyOn(store, "dispatch");
-
-      component.onAddToBatch();
-
-      expect(dispatchSpy).toHaveBeenCalledTimes(2);
-      expect(dispatchSpy).toHaveBeenCalledWith(addToBatchAction());
-      expect(dispatchSpy).toHaveBeenCalledWith(clearSelectionAction());
     });
   });
